Collection interface corrected
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@14620 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtservergql/CGqlCollectionImportControllerComp.cpp
@@ -54,7 +54,7 @@ collectionImport::CSessionStatus CGqlCollectionImportControllerComp::OnBeginColl
 				}
 			}
 
-			if (session.files.size() == sdlImportParamsV1_0.fileList->count()){
+			if (session.files.size() == size_t(sdlImportParamsV1_0.fileList->count())){
 				if (!m_collectionImportControllerCompPtr->BeginCollectionImportSession(session, errorMessage)){
 					retVal.Version_1_0->status = collectionImport::Status::Failed;
 				}
